Membuat Class Child
Membuat Class baru bernama Child, berisi data Child yang ada pada class Employee.
